refactor: separated types from main file and removed useless useMemo
