+++ libs/zephyr-ota-plugin/src/index.ts
@@ -8,12 +8,14 @@ export {
 } from './lib/core/ZephyrOTAService';
 
 // React integration
-export { ZephyrOTAProvider } from './lib/react/ZephyrOTAProvider';
+export {
+  ZephyrOTAProvider,
+  useZephyrOTA,
+} from './lib/react/ZephyrOTAProvider';
 export type {
   ZephyrOTAProviderProps,
   ZephyrOTAContextValue,
 } from './lib/react/ZephyrOTAProvider';
-export { useZephyrOTA } from './lib/react/useZephyrOTA';
 export { useOTAStatus } from './lib/react/useOTAStatus';
 export type { OTAStatusResult } from './lib/react/useOTAStatus';
